Overtime item hours can be float

+++ platform/src/main/java/ru/abondin/hreasy/platform/repo/overtime/OvertimeItemEntry.java
@@ -25,7 +25,7 @@ public class OvertimeItemEntry {
      */
     private LocalDate date;
     private int projectId;
-    private int hours;
+    private float hours;
     private String notes;
 
     private OffsetDateTime createdAt;
